Tenure pagination - temporary changes

+++ FinanceDataMigrationApi/V1/Factories/TransactionFactory.cs
@@ -38,8 +38,8 @@ namespace FinanceDataMigrationApi.V1.Factories
                         TransactionDate = dMTransactionEntityDomain.TransactionDate,
                         TransactionSource = dMTransactionEntityDomain.TransactionSource,
                         TransactionType = dMTransactionEntityDomain.TransactionType.TransactionTypeEnumValue(),
-                        FinancialYear = dMTransactionEntityDomain.FinancialYear,
-                        FinancialMonth = dMTransactionEntityDomain.FinancialMonth
+                        FinancialYear = (short)dMTransactionEntityDomain.FinancialYear,
+                        FinancialMonth = (short)dMTransactionEntityDomain.FinancialMonth
                     };
             }
             catch (Exception e)
